Add range blocklist
